support mipmap

+++ include/editor/Window.hpp
@@ -26,8 +26,8 @@ namespace Guarneri
 		void add_on_resize_evt(void (*on_resize)(int w, int h, void* ud), void* user_data);
 		uint32_t get_fbo() { return FBO; }
 
-		size_t get_width() { return width; }
-		size_t get_height() { return height; }
+		int get_width() { return width; }
+		int get_height() { return height; }
 
 	private:
 		static void glfw_error_callback(int error, const char* description);
@@ -47,8 +47,8 @@ namespace Guarneri
 		bool valid;
 		bool closed;
 
-		size_t width;
-		size_t height;
+		int width;
+		int height;
 
 		uint32_t shader_id;
 		uint32_t FBO;
